[ALLUXIO-2203] Handle NullPointerException in four UFSs

+++ underfs/gcs/src/main/java/alluxio/underfs/gcs/GCSUnderFileSystem.java
@@ -76,9 +76,6 @@ public final class GCSUnderFileSystem extends UnderFileSystem {
   /** The name of the account owner. */
   private final String mAccountOwner;
 
-  /** The GCS user id of the account owner. */
-  private final String mAccountOwnerId;
-
   /** The permission mode that the account owner has to the bucket. */
   private final short mBucketMode;
 
@@ -94,10 +91,11 @@ public final class GCSUnderFileSystem extends UnderFileSystem {
    * Constructs a new instance of {@link GCSUnderFileSystem}.
    *
    * @param uri the {@link AlluxioURI} for this UFS
+   * @return the created {@link GCSUnderFileSystem} instance
    * @throws ServiceException when a connection to GCS could not be created
    */
-  public GCSUnderFileSystem(AlluxioURI uri) throws ServiceException {
-    super(uri);
+  public static GCSUnderFileSystem createInstance(AlluxioURI uri)
+      throws ServiceException {
     String bucketName = uri.getHost();
     Preconditions.checkArgument(Configuration.containsKey(PropertyKey.GCS_ACCESS_KEY),
         "Property " + PropertyKey.GCS_ACCESS_KEY + " is required to connect to GCS");
@@ -106,24 +104,51 @@ public final class GCSUnderFileSystem extends UnderFileSystem {
     GSCredentials googleCredentials = new GSCredentials(
         Configuration.get(PropertyKey.GCS_ACCESS_KEY),
         Configuration.get(PropertyKey.GCS_SECRET_KEY));
-    mBucketName = bucketName;
 
     // TODO(chaomin): maybe add proxy support for GCS.
-    mClient = new GoogleStorageService(googleCredentials);
-    mBucketPrefix = PathUtils.normalizePath(Constants.HEADER_GCS + mBucketName, PATH_SEPARATOR);
+    GoogleStorageService googleStorageService = new GoogleStorageService(googleCredentials);
+    String bucketPrefix = PathUtils.normalizePath(Constants.HEADER_GCS + bucketName,
+        PATH_SEPARATOR);
 
-    mAccountOwnerId = mClient.getAccountOwner().getId();
+    String accountOwnerId = googleStorageService.getAccountOwner().getId();
     // Gets the owner from user-defined static mapping from GCS account id to Alluxio user name.
     String owner = CommonUtils.getValueFromStaticMapping(
-        Configuration.get(PropertyKey.UNDERFS_GCS_OWNER_ID_TO_USERNAME_MAPPING), mAccountOwnerId);
+        Configuration.get(PropertyKey.UNDERFS_GCS_OWNER_ID_TO_USERNAME_MAPPING), accountOwnerId);
     // If there is no user-defined mapping, use the display name.
     if (owner == null) {
-      owner = mClient.getAccountOwner().getDisplayName();
+      owner = googleStorageService.getAccountOwner().getDisplayName();
     }
-    mAccountOwner = owner == null ? mAccountOwnerId : owner;
+    String accountOwner = owner == null ? accountOwnerId : owner;
 
-    GSAccessControlList acl = mClient.getBucketAcl(mBucketName);
-    mBucketMode = GCSUtils.translateBucketAcl(acl, mAccountOwnerId);
+    GSAccessControlList acl = googleStorageService.getBucketAcl(bucketName);
+    short bucketMode = GCSUtils.translateBucketAcl(acl, accountOwnerId);
+
+    return new GCSUnderFileSystem(uri, googleStorageService, bucketName,
+        bucketPrefix, bucketMode, accountOwner);
+  }
+
+  /**
+   * Constructor for {@link GCSUnderFileSystem}.
+   *
+   * @param uri the {@link AlluxioURI} for this UFS
+   * @param googleStorageService the Jets3t GCS client
+   * @param bucketName bucket name of user's configured Alluxio bucket
+   * @param bucketPrefix prefix of the bucket
+   * @param bucketMode the permission mode that the account owner has to the bucket
+   * @param accountOwner the name of the account owner
+   */
+  protected GCSUnderFileSystem(AlluxioURI uri,
+      GoogleStorageService googleStorageService,
+      String bucketName,
+      String bucketPrefix,
+      short bucketMode,
+      String accountOwner) {
+    super(uri);
+    mClient = googleStorageService;
+    mBucketName = bucketName;
+    mBucketPrefix = bucketPrefix;
+    mBucketMode = bucketMode;
+    mAccountOwner = accountOwner;
   }
 
   @Override
@@ -161,15 +186,24 @@ public final class GCSUnderFileSystem extends UnderFileSystem {
   @Override
   public boolean delete(String path, boolean recursive) throws IOException {
     if (!recursive) {
-      if (isFolder(path) && listInternal(path, false).length != 0) {
-        LOG.error("Unable to delete " + path + " because it is a non empty directory. Specify "
-            + "recursive as true in order to delete non empty directories.");
+      String[] children = listInternal(path, false);
+      if (children == null) {
+        LOG.error("Unable to delete {} because listInternal returns null", path);
+        return false;
+      }
+      if (isFolder(path) && children.length != 0) {
+        LOG.error("Unable to delete {} because it is a non empty directory. Specify "
+                + "recursive as true in order to delete non empty directories.", path);
         return false;
       }
       return deleteInternal(path);
     }
     // Get all relevant files
     String[] pathsToDelete = listInternal(path, true);
+    if (pathsToDelete == null) {
+      LOG.error("Unable to delete {} because listInternal returns null", path);
+      return false;
+    }
     for (String pathToDelete : pathsToDelete) {
       // If we fail to deleteInternal one file, stop
       if (!deleteInternal(PathUtils.concatPath(path, pathToDelete))) {
@@ -346,8 +380,13 @@ public final class GCSUnderFileSystem extends UnderFileSystem {
       }
       // Rename each child in the src folder to destination/child
       String [] children = list(src);
-      for (String child: children) {
+      if (children == null) {
+        LOG.error("Failed to list path {}, aborting rename.", src);
+        return false;
+      }
+      for (String child : children) {
         if (!rename(PathUtils.concatPath(src, child), PathUtils.concatPath(dst, child))) {
+          LOG.error("Failed to rename path {}, aborting rename.", child);
           return false;
         }
       }
@@ -599,13 +638,16 @@ public final class GCSUnderFileSystem extends UnderFileSystem {
         for (String commonPrefix : chunk.getCommonPrefixes()) {
           // Remove parent portion of the key
           String child = getChildName(commonPrefix, path);
-          // Remove any portion after the last path delimiter
-          int childNameIndex = child.lastIndexOf(PATH_SEPARATOR);
-          child = childNameIndex != -1 ? child.substring(0, childNameIndex) : child;
-          if (!child.isEmpty() && !children.contains(child)) {
-            // This directory has not been created through Alluxio.
-            mkdirsInternal(commonPrefix);
-            children.add(child);
+
+          if (child != null) {
+            // Remove any portion after the last path delimiter
+            int childNameIndex = child.lastIndexOf(PATH_SEPARATOR);
+            child = childNameIndex != -1 ? child.substring(0, childNameIndex) : child;
+            if (!child.isEmpty() && !children.contains(child)) {
+              // This directory has not been created through Alluxio.
+              mkdirsInternal(commonPrefix);
+              children.add(child);
+            }
           }
         }
         done = chunk.isListingComplete();
